requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.2
MySQL-python==1.2.3
Flask-SQLAlchemy==2.1
SQLAlchemy==1.0.12
jsonify==0.5
Flask-Login==0.4.0
Flask-OpenID==1.2.5
PyYAML==3.12
